mysite/requirements.txt
aioredis==1.3.1
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
biopython==1.79
bleach==5.0.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
channels==2.4.0
channels-redis==3.4.1
chardet==3.0.4
click==8.1.3
colour==0.1.5
constantly==15.1.0
cryptography==38.0.1
daphne==2.5.0
dash==2.6.2
dash-bio==1.0.2
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.13
Django==3.2.0
django-cookies-samesite==0.9.0
django-cors-headers==4.3.1
django-rest-authemail==2.1.7
django-plotly-dash==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dpd-components==0.1.0
dpd-static-support==0.0.5
drf-yasg==1.21.8
entrypoints==0.3
Flask==2.1.3
Flask-Compress==1.13
GEOparse==2.0.3
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==5.0.0
incremental==22.10.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.4
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.22.3
packaging==21.0
pandas==1.5.0
pandocfilters==1.4.3
paramiko==2.9.2
ParmEd==3.4.3
parso==0.8.1
periodictable==1.6.1
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
plotly==5.10.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
redis==4.3.4
requests==2.25.0
scikit-learn==1.1.2
scipy==1.9.2
Send2Trash==1.5.0
service-identity==21.1.0
simplejson==3.17.5
six==1.15.0
sqlparse==0.4.1
tenacity==8.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.1
traitlets==5.0.5
Twisted==22.8.0
txaio==22.2.1
typing-extensions==4.4.0
ua-parser==0.10.0
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wrapt==1.14.1
zipp==3.9.0
zope.interface==5.5.0
